This closes wso2/product-iots#1677
After uploading an application to the publisher portal it displays applications and app count correctly on the publisher portal as well as store portal.
Thereafter, If we delete any application from the publisher portal it will delete the application correctly but, it hides some applications which already exist in the publisher portal.
In addition to that, after deleting an application it doesn't display recent applications correctly on the store portal.

This fix change the app deletion logic. Instead of deleting app collection delete only application resource.