This closes wso2/product-iots#1698
When installing an iOS application through store portal in a distributed IoTS setup, iOS device tries to reach to the IoT server through an inaccessible port. In order to fix this issue, it is required to use correct port when the device tries to reach the server. Hence, instead of using carbon properties used System properties to get correct port.